requirements.txt
# Web framework
fastapi==0.111.0
uvicorn==0.29.0

# Database
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.7

# Data validation
pydantic==2.7.1

# AI data generation
anthropic==0.25.8
faker==25.0.1
python-dotenv==1.0.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-html==4.1.1
pytest-playwright==0.5.0
httpx==0.27.0
